Parameter 4 of constructor in org.intecbrussel.onlinecoursespringproject.service.EnrollmentServiceImpl required a bean of type 'org.intecbrussel.onlinecoursespringproject.repository.CustomEnrollmentRepository' that could not be found.
Action:

Consider defining a bean of type 'org.intecbrussel.onlinecoursespringproject.repository.CustomEnrollmentRepository' in your configuration.

+++ src/main/java/org/intecbrussel/onlinecoursespringproject/repository/EnrollmentRepository.java
@@ -1,10 +1,12 @@
 package org.intecbrussel.onlinecoursespringproject.repository;
 
 import org.intecbrussel.onlinecoursespringproject.model.Enrollment;
+import org.intecbrussel.onlinecoursespringproject.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface EnrollmentRepository extends JpaRepository<Enrollment, Integer> {
-    ;
+public interface EnrollmentRepository extends JpaRepository<Enrollment, Long> {
 }
